+++ demisto_sdk/commands/common/errors.py
@@ -2,6 +2,7 @@ from distutils.version import LooseVersion
 from typing import Any, Dict, List, Optional
 
 import decorator
+from requests import Response
 
 from demisto_sdk.commands.common.constants import (BETA_INTEGRATION_DISCLAIMER,
                                                    CONF_PATH,
@@ -51,6 +52,7 @@ ERROR_CODE = {
     'entity_name_contains_excluded_word': {'code': 'BA111', 'ui_applicable': False, 'related_field': ''},
     "spaces_in_the_end_of_id": {'code': "BA112", 'ui_applicable': False, 'related_field': 'id'},
     "spaces_in_the_end_of_name": {'code': "BA113", 'ui_applicable': False, 'related_field': 'name'},
+    "changed_pack_name": {'code': "BA114", 'ui_applicable': False, 'related_field': 'name'},
 
     # BC - Backward Compatible
     "breaking_backwards_subtype": {'code': "BC100", 'ui_applicable': False, 'related_field': 'subtype'},
@@ -281,6 +283,7 @@ ERROR_CODE = {
     "metadata_url_invalid": {'code': "PA127", 'ui_applicable': False, 'related_field': ''},
     "required_pack_file_does_not_exist": {'code': "PA128", 'ui_applicable': False, 'related_field': ''},
     "pack_metadata_missing_categories": {'code': "PA129", 'ui_applicable': False, 'related_field': ''},
+    "wrong_version_format": {'code': "PA130", 'ui_applicable': False, 'related_field': ''},
 
     # PB - Playbooks
     "playbook_cant_have_rolename": {'code': "PB100", 'ui_applicable': True, 'related_field': 'rolename'},
@@ -1646,14 +1649,20 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
-    def invalid_readme_image_error(path, error_type):
-        return 'Error in readme image:\n' + {
-            'pack_readme_relative_error': Errors.pack_readme_image_relative_path_error,
-            'general_readme_relative_error': Errors.invalid_readme_image_relative_path_error,
-            'general_readme_absolute_error': Errors.invalid_readme_image_absolute_path_error,
-            'branch_name_readme_absolute_error': Errors.branch_name_in_readme_image_absolute_path_error,
-            'insert_image_link_error': Errors.invalid_readme_insert_image_link_error
-        }.get(error_type, lambda x: f'Something went wrong when testing {x}')(path)
+    def invalid_readme_image_error(path: str, error_type: str, response: Optional[Response] = None):
+        error = 'Error in readme image: '
+        if response is not None:
+            error += f'got HTTP response code {response.status_code}'
+            error += f', reason = {response.reason}' if response.reason else " "
+
+        error_body = {'pack_readme_relative_error': Errors.pack_readme_image_relative_path_error,
+                      'general_readme_relative_error': Errors.invalid_readme_image_relative_path_error,
+                      'general_readme_absolute_error': Errors.invalid_readme_image_absolute_path_error,
+                      'branch_name_readme_absolute_error': Errors.branch_name_in_readme_image_absolute_path_error,
+                      'insert_image_link_error': Errors.invalid_readme_insert_image_link_error} \
+            .get(error_type, lambda x: f'Unexpected error when testing {x}')(path)
+
+        return error + f"\n{error_body}"
 
     @staticmethod
     @error_code_decorator
@@ -2109,3 +2118,15 @@ class Errors:
     @error_code_decorator
     def invalid_yml_file(error):
         return f'There is problem with the yml file. The error: {error}'
+
+    @staticmethod
+    @error_code_decorator
+    def changed_pack_name(original_name):
+        return f'Pack folder names cannot be changed, please rename it back to {original_name}.' \
+               f' If you wish to rename the pack, you can edit the name field in pack_metadata.json,' \
+               f' and the pack will be shown in the Marketplace accordingly.'
+
+    @staticmethod
+    @error_code_decorator
+    def wrong_version_format():
+        return 'Pack metadata version format is not valid. Please fill in a valid format (example: 0.0.0)'