+++ demisto_sdk/commands/common/tests/update_id_set_test.py
@@ -8,13 +8,16 @@ from pathlib import Path
 
 import pytest
 
+import demisto_sdk.commands.common.tools as tools
 import demisto_sdk.commands.common.update_id_set as uis
 from demisto_sdk.commands.common.constants import (DEFAULT_JOB_FROM_VERSION,
-                                                   JOBS_DIR, FileType)
+                                                   JOBS_DIR, FileType,
+                                                   MarketplaceVersions)
 from demisto_sdk.commands.common.legacy_git_tools import git_path
 from demisto_sdk.commands.common.update_id_set import (
-    does_dict_have_alternative_key, find_duplicates, get_classifier_data,
-    get_dashboard_data, get_fields_by_script_argument,
+    add_item_to_exclusion_dict, does_dict_have_alternative_key,
+    find_duplicates, get_classifier_data, get_dashboard_data,
+    get_fields_by_script_argument,
     get_filters_and_transformers_from_complex_value,
     get_filters_and_transformers_from_playbook, get_general_data,
     get_generic_field_data, get_generic_module_data, get_generic_type_data,
@@ -81,8 +84,8 @@ class TestPacksMetadata:
         Then
             - parsing all the data from file successfully
         """
-        mocker.patch.object(uis, 'get_pack_name', return_value='Pack1')
-        mocker.patch.object(uis, 'get_mp_types_from_metadata_by_item', return_value=['xsoar', 'marketplacev2'])
+        mocker.patch.object(tools, 'get_pack_name', return_value='Pack1')
+        mocker.patch.object(tools, 'get_mp_types_from_metadata_by_item', return_value=['xsoar', 'marketplacev2'])
 
         pack = repo.create_pack("Pack1")
         pack.pack_metadata.write_json(metadata_file_content)
@@ -285,6 +288,7 @@ class TestDuplicates:
                 'fromVersion': '5.0.0',
                 'kind': 'Details',
                 'path': 'Layouts/layout-details-urlrep.json',
+                'marketplaces': ['xsoar'],
                 'pack': first_pack,
                 'source': first_source
             }
@@ -295,6 +299,7 @@ class TestDuplicates:
                 'typeID': 'urlRep',
                 'kind': 'Details',
                 'path': 'Layouts/layout-details-urlrep2.json',
+                'marketplaces': ['xsoar'],
                 'pack': second_pack,
                 'source': second_source
             }
@@ -307,7 +312,7 @@ class TestDuplicates:
 
         # Check for duplicates for a new id-set,
         # In this case all the examples above should be considered as duplicates
-        assert has_duplicate(id_set['Layouts'], 'urlRep', 'Layouts', False, is_create_new=True),\
+        assert has_duplicate(id_set['Layouts'], 'urlRep', 'Layouts', False, is_create_new=True), \
             "if it's a new pack it is always a duplicate"
 
     @staticmethod
@@ -344,6 +349,41 @@ class TestDuplicates:
         has_duplicates = has_duplicate(id_set['Layouts'], 'urlRep', 'Layouts', False)
         assert has_duplicates is False
 
+    @staticmethod
+    def test_similar_items_on_different_marketplaces():
+        """
+        Given
+            - id_set with two integrations with the same IDs with different marketplaces.
+
+        When
+            - checking for duplicate
+
+        Then
+            - Ensure duplicates not found
+        """
+
+        integrations = [
+            {
+                'Integ': {
+                    'id': 'Integ',
+                    'name': 'Integ',
+                    'pack': 'Reut',
+                    'marketplaces': ['xsoar'],
+                }
+            },
+            {
+                'Integ': {
+                    'id': 'Integ',
+                    'name': 'Integ',
+                    'pack': 'Cactus',
+                    'marketplaces': ['marketplacev2'],
+                }
+
+            },
+        ]
+        has_duplicates = has_duplicate(integrations, 'Integ', 'integrations', print_logs=False)
+        assert not has_duplicates
+
 
 class TestIntegrations:
     INTEGRATION_DATA = {
@@ -352,6 +392,7 @@ class TestIntegrations:
             "file_path": TESTS_DIR + "/test_files/DummyPack/Integrations/DummyIntegration/DummyIntegration.yml",
             "fromversion": "4.1.0",
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "docker_image": "demisto/python3:3.7.4.977",
             "commands": ['xdr-get-incidents',
                          'xdr-get-incident-extra-data',
@@ -372,10 +413,11 @@ class TestIntegrations:
             "incident_types": "dummy-incident-type",
             "indicator_fields": "CommonTypes",
             "indicator_types": "CommonTypes",
+            "is_fetch": True,
             "mappers": [
                 "dummy-mapper-in",
                 "dummy-mapper-out"
-            ]
+            ],
         }
     }
 
@@ -385,6 +427,7 @@ class TestIntegrations:
             "file_path": TESTS_DIR + "/test_files/DummyPack/Integrations/integration-DummyIntegration.yml",
             "fromversion": "4.1.0",
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "commands": ['xdr-get-incidents',
                          'xdr-get-incident-extra-data',
                          'xdr-update-incident',
@@ -404,17 +447,18 @@ class TestIntegrations:
             "incident_types": "dummy-incident-type",
             "indicator_fields": "CommonTypes",
             "indicator_types": "CommonTypes",
+            "is_fetch": True,
             "mappers": [
                 "dummy-mapper-in",
                 "dummy-mapper-out"
-            ]
+            ],
         }
     }
 
     def test_process_integration__sanity(self, mocker):
         """
         Given
-            - A valid script package folder located at Packs/DummyPack/Scripts/DummyScript.
+            - A valid script package folder located at Packs/DummyPack/Integrations/DummyIntegration.
 
         When
             - parsing script files
@@ -425,14 +469,14 @@ class TestIntegrations:
         non_unified_file_path = os.path.join(TESTS_DIR, 'test_files',
                                              'DummyPack', 'Integrations', 'DummyIntegration')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
-        res = process_integration(non_unified_file_path, True)
+        res, _ = process_integration(non_unified_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=True)
         assert len(res) == 1
         non_unified_integration_data = res[0]
 
         unified_file_path = os.path.join(TESTS_DIR, 'test_files',
                                          'DummyPack', 'Integrations', 'integration-DummyIntegration.yml')
 
-        res = process_integration(unified_file_path, True)
+        res, _ = process_integration(unified_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=True)
         assert len(res) == 1
         unified_integration_data = res[0]
 
@@ -442,12 +486,12 @@ class TestIntegrations:
         ]
 
         for returned, constant in test_pairs:
-            assert IsEqualFunctions.is_lists_equal(list(returned.keys()), list(constant.keys()))
+            assert list(returned.keys()) == list(constant.keys())
 
             const_data = constant.get('Dummy Integration')
             returned_data = returned.get('Dummy Integration')
 
-            assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
+            assert IsEqualFunctions.is_dicts_equal(returned_data, const_data, lists_as_sets=True)
 
     @staticmethod
     def test_process_integration__exception(mocker):
@@ -464,7 +508,7 @@ class TestIntegrations:
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
         with pytest.raises(Exception):
-            process_integration(test_file_path, True)
+            process_integration(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=True)
 
     @staticmethod
     def test_process_integration__marketplace_mismatch(mocker):
@@ -480,7 +524,7 @@ class TestIntegrations:
         """
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
-        res = process_integration(test_file_path, print_logs=False)
+        res, _ = process_integration(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=False)
         assert res == []
 
 
@@ -490,6 +534,7 @@ class TestScripts:
             "name": "DummyScript",
             "file_path": TESTS_DIR + "/test_files/DummyPack/Scripts/DummyScript2.yml",
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "fromversion": "5.0.0",
             "docker_image": "demisto/python3:3.7.3.286",
             "tests": [
@@ -503,6 +548,7 @@ class TestScripts:
             "name": "DummyScript",
             "file_path": TESTS_DIR + '/test_files/alternative_meta_fields/Script-top_level_alternative_fields.yml',
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "fromversion": "5.0.0",
             "docker_image": "demisto/python3:3.7.3.286",
             "tests": [
@@ -517,6 +563,7 @@ class TestScripts:
             "name": "DummyScript",
             "file_path": TESTS_DIR + '/test_files/alternative_meta_fields/Script-second_level_alternative_fields.yml',
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "fromversion": "5.0.0",
             "docker_image": "demisto/python3:3.7.3.286",
             "tests": [
@@ -529,10 +576,11 @@ class TestScripts:
     PACK_SCRIPT_DATA = {
         "DummyScript": {
             "name": "DummyScript",
+            "file_path": TESTS_DIR + "/test_files/Packs/DummyPack/Scripts/DummyScript/DummyScript.yml",
             "source": ['github.com', 'demisto', 'demisto-sdk'],
+            "marketplaces": ["xsoar"],
             "docker_image": "demisto/python3:3.8.2.6981",
             "pack": "DummyPack",
-            "file_path": TESTS_DIR + "/test_files/Packs/DummyPack/Scripts/DummyScript/DummyScript.yml",
         }
     }
 
@@ -542,9 +590,9 @@ class TestScripts:
         Test for getting the script data
         """
         file_path = TESTS_DIR + '/test_files/DummyPack/Scripts/DummyScript2.yml'
-        data = get_script_data(file_path)
+        data = get_script_data(file_path, packs={'DummyPack': {}})
 
-        assert IsEqualFunctions.is_lists_equal(list(data.keys()), list(TestScripts.SCRIPT_DATA.keys()))
+        assert list(data.keys()) == list(TestScripts.SCRIPT_DATA.keys())
 
         const_data = TestScripts.SCRIPT_DATA.get('DummyScript')
         returned_data = data.get('DummyScript')
@@ -565,9 +613,9 @@ class TestScripts:
             - making sure the entry has a field called "has_alternative_meta" set to True
         """
         file_path = TESTS_DIR + '/test_files/alternative_meta_fields/Script-top_level_alternative_fields.yml'
-        data = get_script_data(file_path)
+        data = get_script_data(file_path, packs={'DummyPack': {}})
 
-        assert IsEqualFunctions.is_lists_equal(list(data.keys()), list(TestScripts.SCRIPT_DATA_ALTERNATIVE_TOP_LEVEL.keys()))
+        assert list(data.keys()) == list(TestScripts.SCRIPT_DATA_ALTERNATIVE_TOP_LEVEL.keys())
 
         const_data = TestScripts.SCRIPT_DATA_ALTERNATIVE_TOP_LEVEL.get('DummyScript')
         returned_data = data.get('DummyScript')
@@ -588,10 +636,9 @@ class TestScripts:
             - making sure the entry has a field called "has_alternative_meta" set to True
         """
         file_path = TESTS_DIR + '/test_files/alternative_meta_fields/Script-second_level_alternative_fields.yml'
-        data = get_script_data(file_path)
+        data = get_script_data(file_path, packs={'DummyPack': {}})
 
-        assert IsEqualFunctions.is_lists_equal(list(data.keys()),
-                                               list(TestScripts.SCRIPT_DATA_ALTERNATIVE_SECOND_LEVEL.keys()))
+        assert (list(data.keys()) == list(TestScripts.SCRIPT_DATA_ALTERNATIVE_SECOND_LEVEL.keys()))
 
         const_data = TestScripts.SCRIPT_DATA_ALTERNATIVE_SECOND_LEVEL.get('DummyScript')
         returned_data = data.get('DummyScript')
@@ -613,11 +660,11 @@ class TestScripts:
         test_file_path = os.path.join(TESTS_DIR, 'test_files',
                                       'Packs', 'DummyPack', 'Scripts', 'DummyScript')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
-        res = process_script(test_file_path, True)
+        res, _ = process_script(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=True)
         assert len(res) == 1
         data = res[0]
 
-        assert IsEqualFunctions.is_lists_equal(list(data.keys()), list(TestScripts.PACK_SCRIPT_DATA.keys()))
+        assert list(data.keys()) == list(TestScripts.PACK_SCRIPT_DATA.keys())
 
         const_data = TestScripts.PACK_SCRIPT_DATA.get('DummyScript')
         returned_data = data.get('DummyScript')
@@ -639,7 +686,7 @@ class TestScripts:
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
-        res = process_script(test_file_path, print_logs=False)
+        res, _ = process_script(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=False)
         assert res == []
 
     @staticmethod
@@ -658,7 +705,7 @@ class TestScripts:
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         with pytest.raises(Exception):
-            process_script(test_file_path, True)
+            process_script(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, print_logs=True)
 
 
 class TestPlaybooks:
@@ -667,6 +714,7 @@ class TestPlaybooks:
         "file_path": TESTS_DIR + "/test_files/DummyPack/Playbooks/DummyPlaybook.yml",
         "source": ['github.com', 'demisto', 'demisto-sdk'],
         "fromversion": "4.5.0",
+        "marketplaces": ["xsoar"],
         "filters": ["isEqualString"],
         "transformers": ["uniq"],
         "implementing_scripts": [
@@ -674,8 +722,8 @@ class TestPlaybooks:
             "StopScheduledTask",
         ],
         "implementing_playbooks": [
-            "Calculate Severity - Standard",
             "Palo Alto Networks - Malware Remediation",
+            "Calculate Severity - Standard"
         ],
         "command_to_integration": {
             "xdr-update-incident": "",
@@ -686,6 +734,7 @@ class TestPlaybooks:
         ],
         "skippable_tasks": [
             "StopScheduledTask",
+            "Palo Alto Networks - Malware Remediation",
             "autofocus-sample-analysis"
         ]
     }
@@ -695,11 +744,12 @@ class TestPlaybooks:
         "file_path": TESTS_DIR + "/test_files/alternative_meta_fields/Playbook-top_level_alternative_fields.yml",
         "source": ['github.com', 'demisto', 'demisto-sdk'],
         "fromversion": "4.5.0",
+        "marketplaces": ["xsoar"],
         "filters": ["isEqualString"],
         "transformers": ["uniq"],
         "implementing_scripts": [
-            "XDRSyncScript",
             "StopScheduledTask",
+            "XDRSyncScript"
         ],
         "implementing_playbooks": [
             "Calculate Severity - Standard",
@@ -714,6 +764,7 @@ class TestPlaybooks:
         ],
         "skippable_tasks": [
             "StopScheduledTask",
+            "Palo Alto Networks - Malware Remediation",
             "autofocus-sample-analysis"
         ],
         "has_alternative_meta": True
@@ -723,6 +774,7 @@ class TestPlaybooks:
         "name": "Dummy Playbook",
         "file_path": TESTS_DIR + "/test_files/alternative_meta_fields/Playbook-second_level_alternative_fields.yml",
         "source": ['github.com', 'demisto', 'demisto-sdk'],
+        "marketplaces": ["xsoar"],
         "fromversion": "4.5.0",
         "filters": ["isEqualString"],
         "transformers": ["uniq"],
@@ -743,6 +795,7 @@ class TestPlaybooks:
         ],
         "skippable_tasks": [
             "StopScheduledTask",
+            "Palo Alto Networks - Malware Remediation",
             "autofocus-sample-analysis"
         ],
         "has_alternative_meta": True
@@ -754,9 +807,8 @@ class TestPlaybooks:
         Test for getting the playbook data
         """
         file_path = TESTS_DIR + '/test_files/DummyPack/Playbooks/DummyPlaybook.yml'
-        data = get_playbook_data(file_path)['Dummy Playbook']
-
-        assert IsEqualFunctions.is_dicts_equal(data, TestPlaybooks.PLAYBOOK_DATA)
+        data = get_playbook_data(file_path, packs={'DummyPack': {}})['Dummy Playbook']
+        assert IsEqualFunctions.is_dicts_equal(data, TestPlaybooks.PLAYBOOK_DATA, lists_as_sets=True)
 
     @staticmethod
     def test_get_playbook_data_with_alternative_fields_top_level():
@@ -772,9 +824,11 @@ class TestPlaybooks:
             - making sure the entry has a field called "has_alternative_meta" set to True
         """
         file_path = TESTS_DIR + '/test_files/alternative_meta_fields/Playbook-top_level_alternative_fields.yml'
-        data = get_playbook_data(file_path)['Dummy Playbook']
+        data = get_playbook_data(file_path, packs={'DummyPack': {}})['Dummy Playbook']
 
-        assert IsEqualFunctions.is_dicts_equal(data, TestPlaybooks.PLAYBOOK_DATA_ALTERNATIVE_FIELDS_TOP_LEVEL)
+        assert IsEqualFunctions.is_dicts_equal(data,
+                                               TestPlaybooks.PLAYBOOK_DATA_ALTERNATIVE_FIELDS_TOP_LEVEL,
+                                               lists_as_sets=True)
 
     @staticmethod
     def test_get_playbook_data_with_alternative_fields_second_level():
@@ -790,9 +844,10 @@ class TestPlaybooks:
             - making sure the entry has a field called "has_alternative_meta" set to True
         """
         file_path = TESTS_DIR + '/test_files/alternative_meta_fields/Playbook-second_level_alternative_fields.yml'
-        data = get_playbook_data(file_path)['Dummy Playbook']
-
-        assert IsEqualFunctions.is_dicts_equal(data, TestPlaybooks.PLAYBOOK_DATA_ALTERNATIVE_FIELDS_SECOND_LEVEL)
+        data = get_playbook_data(file_path, packs={'DummyPack': {}})['Dummy Playbook']
+        assert IsEqualFunctions.is_dicts_equal(data,
+                                               TestPlaybooks.PLAYBOOK_DATA_ALTERNATIVE_FIELDS_SECOND_LEVEL,
+                                               lists_as_sets=True)
 
     @staticmethod
     def test_get_playbook_data_2():
@@ -807,7 +862,7 @@ class TestPlaybooks:
             - parsing all the data from file successfully
         """
         test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-with-incident-fields.yml'
-        result = get_playbook_data(test_dir)
+        result = get_playbook_data(test_dir, packs={'DummyPack': {}})
         result = result.get('Arcsight - Get events related to the Case')
         assert 'name' in result.keys()
         assert 'file_path' in result.keys()
@@ -833,7 +888,7 @@ class TestPlaybooks:
         test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-no-incident-fields.yml'
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        result = get_playbook_data(test_dir)
+        result = get_playbook_data(test_dir, packs={'DummyPack': {}})
         result = result.get('Arcsight - Get events related to the Case')
         assert 'name' in result.keys()
         assert 'file_path' in result.keys()
@@ -858,7 +913,7 @@ class TestPlaybooks:
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'playbook.yml')
         with pytest.raises(Exception):
-            process_general_items(test_file_path, True, (FileType.PLAYBOOK,), get_playbook_data)
+            process_general_items(test_file_path, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, (FileType.PLAYBOOK,), get_playbook_data)
 
     @staticmethod
     def test_get_playbook_data_bad_graph():
@@ -875,7 +930,7 @@ class TestPlaybooks:
             - parsing flow graph from file successfully (only tasks 0 and 1 will be in the graph)
         """
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'playbook-invalid-bad-graph.yml')
-        result = get_playbook_data(test_file_path)
+        result = get_playbook_data(test_file_path, packs={'DummyPack': {}})
         playbook_data = result.get('InvalidPlaybook-BadGraph', {})
         assert playbook_data.get('name') == 'InvalidPlaybook-BadGraph'
         assert playbook_data.get('command_to_integration', {}).get('ip') == ''
@@ -897,7 +952,7 @@ class TestPlaybooks:
             - parsing flow graph from file successfully (no actual tasks will be in the graph)
         """
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'playbook-invalid-bad-graph_2.yml')
-        result = get_playbook_data(test_file_path)
+        result = get_playbook_data(test_file_path, packs={'DummyPack': {}})
         playbook_data = result.get('InvalidPlaybook-BadGraph', {})
         assert playbook_data.get('name') == 'InvalidPlaybook-BadGraph'
         assert playbook_data.get('command_to_integration', {}).get('ip') == ''
@@ -1039,7 +1094,7 @@ class TestLayouts:
                                  'test_data', 'layout-to-test.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_general_items(test_file, True, (FileType.LAYOUT,), get_layout_data)
+        res, _ = process_general_items(test_file, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, (FileType.LAYOUT,), get_layout_data)
         assert len(res) == 1
         result = res[0]
         result = result.get('urlRep')
@@ -1068,7 +1123,7 @@ class TestLayouts:
                                  'test_data', 'layout-to-test-no-types-fields.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_general_items(test_file, False, (FileType.LAYOUT,), get_layout_data)
+        res, _ = process_general_items(test_file, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, False, (FileType.LAYOUT,), get_layout_data)
         assert len(res) == 1
         result = res[0]
         result = result.get('urlRep')
@@ -1095,7 +1150,8 @@ class TestLayouts:
                                  'test_data', 'layoutscontainer-to-test.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_general_items(test_file, True, (FileType.LAYOUTS_CONTAINER,), get_layoutscontainer_data)
+        res, _ = process_general_items(test_file, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value,
+                                       True, (FileType.LAYOUTS_CONTAINER,), get_layoutscontainer_data)
         assert len(res) == 1
         result = res[0]
         result = result.get('layouts_container_test')
@@ -1126,7 +1182,7 @@ class TestIncidentFields:
                                 'test_data', 'incidentfield-to-test.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_incident_fields(test_dir, True, [])
+        res, _ = process_incident_fields(test_dir, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, [])
         assert len(res) == 1
         result = res[0]
         result = result.get('incidentfield-test')
@@ -1154,7 +1210,7 @@ class TestIncidentFields:
                                 'incidentfield-top_level_alternative_fields.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_incident_fields(test_dir, True, [])
+        res, _ = process_incident_fields(test_dir, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, [])
         assert len(res) == 1
         result = res[0]
         result = result.get('incidentfield_upload_id')
@@ -1181,7 +1237,7 @@ class TestIncidentFields:
                                 'test_data', 'incidentfield-to-test-no-types_scripts.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_incident_fields(test_dir, True, [])
+        res, _ = process_incident_fields(test_dir, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, [])
         assert len(res) == 1
         result = res[0]
         result = result.get('incidentfield-test')
@@ -1205,7 +1261,7 @@ class TestIncidentFields:
         test_dir = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                 'test_data', 'incidentfield-to-test-no-types_scripts.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
-        res = process_incident_fields(test_dir, True, [])
+        res, _ = process_incident_fields(test_dir, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, [])
         assert res == []
 
 
@@ -1443,6 +1499,7 @@ class TestMappers:
             "feed": False,
             "fromVersion": "6.0.0",
             "id": "ServiceNow-outgoing-mapper",
+            "marketplaces": ["xsoar"],
             "mapping": {
                 "ServiceNow Ticket": {
                     "dontMapEventToLabels": False,
@@ -1811,6 +1868,7 @@ class TestDashboard:
             }
         ],
         "name": "dummy_dashboard",
+        "marketplaces": ["xsoar"],
         "fromVersion": "6.0.0",
     }
 
@@ -1832,6 +1890,7 @@ class TestDashboard:
             }
         ],
         "name": "dummy_dashboard",
+        "marketplaces": ["xsoar"],
         "fromVersion": "6.0.0",
     }
 
@@ -1895,6 +1954,7 @@ class TestReport:
         "modified": "2020-09-23T07:54:57.783240299Z",
         "startDate": "0001-01-01T00:00:00Z",
         "name": "dummy_report",
+        "marketplaces": ["xsoar"],
         "dashboard": {
             "id": "dummy_report",
             "version": 0,
@@ -1919,6 +1979,7 @@ class TestReport:
     REPORT_NO_SCRIPT = {
         "id": "dummy_report",
         "name": "dummy_report",
+        "marketplaces": ["xsoar"],
         "dashboard": {
             "id": "dummy_report",
             "name": "dummy_report",
@@ -2005,7 +2066,7 @@ class TestGenericFunctions:
                                  'test_data', 'classifier-to-test.json')
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
-        res = process_general_items(test_file, True, (FileType.CLASSIFIER,), get_classifier_data)
+        res, _ = process_general_items(test_file, {'': {}}, MarketplaceVersions.XSOAR.value, True, (FileType.CLASSIFIER,), get_classifier_data)
         assert len(res) == 1
         result = res[0]
         result = result.get('dummy classifier')
@@ -2032,7 +2093,7 @@ class TestGenericFunctions:
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'classifier-to-test.json')
-        res = process_general_items(test_file, True, (FileType.CLASSIFIER,), get_classifier_data)
+        res, _ = process_general_items(test_file, {'DummyPack': {}}, MarketplaceVersions.XSOAR.value, True, (FileType.CLASSIFIER,), get_classifier_data)
         assert res == []
 
     @staticmethod
@@ -2554,7 +2615,7 @@ class TestJob:
         job_details = 'job details'
         mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         job = pack.create_job(is_feed, details=job_details)
-        res = process_jobs(job.path, print_logs)
+        res = process_jobs(job.path, {pack.name: {}}, MarketplaceVersions.XSOAR.value, print_logs)
 
         captured = capsys.readouterr()
         assert len(res) == 1
@@ -2593,7 +2654,7 @@ class TestJob:
         job = pack.create_job(is_feed)
         job_path = Path(job.path)
         new_path = job_path.rename(job_path.with_suffix('.yml'))
-        res = process_jobs(str(new_path), False)
+        res = process_jobs(str(new_path), {pack.name: {}}, MarketplaceVersions.XSOAR.value, False)
         assert not res
 
     @staticmethod
@@ -2618,7 +2679,7 @@ class TestJob:
 
         job_json_path.rename(job_json_path_as_yml)
         with pytest.raises(FileNotFoundError):
-            assert not process_jobs(str(job_json_path), print_logs)
+            assert not process_jobs(str(job_json_path), {pack.name: {}}, MarketplaceVersions.XSOAR.value, print_logs)
         assert f"failed to process job {job_json_path}" in capsys.readouterr().out
 
 
@@ -2744,7 +2805,8 @@ def test_merged_id_sets_with_duplicates(caplog):
         'playbooks': [
             {
                 'playbook_foo1': {
-                    'name': 'playbook_foo1'
+                    'name': 'playbook_foo1',
+                    'marketplaces': 'xsoar',
                 }
             }
         ],
@@ -2752,7 +2814,8 @@ def test_merged_id_sets_with_duplicates(caplog):
             {
                 'ScriptFoo': {
                     'name': 'ScriptFoo',
-                    'pack': 'ScriptFoo1'
+                    'pack': 'ScriptFoo1',
+                    'marketplaces': 'xsoar',
                 }
             }
         ]
@@ -2762,7 +2825,8 @@ def test_merged_id_sets_with_duplicates(caplog):
         'playbooks': [
             {
                 'playbook_foo2': {
-                    'name': 'playbook_foo2'
+                    'name': 'playbook_foo2',
+                    'marketplaces': 'xsoar',
                 }
             }
         ],
@@ -2770,7 +2834,8 @@ def test_merged_id_sets_with_duplicates(caplog):
             {
                 'ScriptFoo': {
                     'name': 'ScriptFoo',
-                    'pack': 'ScriptFoo2'
+                    'pack': 'ScriptFoo2',
+                    'marketplaces': 'xsoar',
                 }
             }
         ]
@@ -2805,7 +2870,8 @@ def test_merged_id_sets_with_legal_duplicates(caplog):
                 'playbook_foo1': {
                     'name': 'playbook_foo1',
                     'pack': 'foo_1',
-                    "source": ['github.com', 'demisto', 'repo1'],
+                    'marketplaces': 'xsoar',
+                    'source': ['github.com', 'demisto', 'repo1'],
                 }
             }
         ],
@@ -2814,7 +2880,8 @@ def test_merged_id_sets_with_legal_duplicates(caplog):
                 'Script_Foo1': {
                     'name': 'ScriptFoo',
                     'pack': 'foo_1',
-                    "source": ['github.com', 'demisto', 'repo1'],
+                    'marketplaces': 'xsoar',
+                    'source': ['github.com', 'demisto', 'repo1'],
                 }
             }
         ]
@@ -2826,7 +2893,8 @@ def test_merged_id_sets_with_legal_duplicates(caplog):
                 'playbook_foo1': {
                     'name': 'playbook_foo1',
                     'pack': 'foo_1',
-                    "source": ['github.com', 'demisto', 'repo2'],
+                    'marketplaces': 'xsoar',
+                    'source': ['github.com', 'demisto', 'repo2'],
                 }
             }
         ],
@@ -2871,27 +2939,6 @@ def test_get_filters_and_transformers_from_complex_value():
     assert 'StringContainsArray' in filters
 
 
-def test_should_skip_item_by_mp(mocker):
-    """
-    Given
-    - path of content item, the current marketplace this id set is generated for.
-    When
-    - when creating the id set, checking a content item that is 'xsoar only' if it should be part of the mpV2
-     id set.
-    Then
-    - return True since this item should be skipped.
-
-    """
-    import demisto_sdk.commands.common.update_id_set as uis
-    mocker.patch.object(uis, 'get_mp_types_from_metadata_by_item', return_value=['xsoar'])
-    pack_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackXsoarMPOnly')
-    script_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackScriptIsXsoarOnly', 'Scripts', 'DummyScript')
-    res1 = should_skip_item_by_mp(pack_path, 'mpv2')
-    res2 = should_skip_item_by_mp(script_path, 'mpv2')
-    assert res1
-    assert res2
-
-
 TEST_DICTS = [
     (
         {
@@ -2934,3 +2981,83 @@ TEST_DICTS = [
 def test_does_dict_have_alternative_key(dict_to_test, expected_result):
     result = does_dict_have_alternative_key(dict_to_test)
     assert result == expected_result
+
+
+def test_should_skip_item_by_mp(mocker):
+    """
+    Given
+    - path of content item, the current marketplace this id set is generated for.
+    When
+    - when creating the id set, checking a content item that is 'xsoar only' if it should be part of the mpV2
+     id set.
+    Then
+    - return True since this item should be skipped.
+
+    """
+    import demisto_sdk.commands.common.tools as tools
+    mocker.patch.object(tools, 'get_mp_types_from_metadata_by_item', return_value=['xsoar'])
+    pack_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackXsoarMPOnly', 'pack_metadata.json')
+    script_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackScriptIsXsoarOnly', 'Scripts', 'DummyScript.yml')
+    res1 = should_skip_item_by_mp(pack_path, 'mpv2', {})
+    res2 = should_skip_item_by_mp(script_path, 'mpv2', {})
+    assert res1
+    assert res2
+
+
+def test_should_skip_item_by_mp_no_update_excluded_dict(mocker):
+    """
+    Given
+        - path of a pack which related only to xsoar marketplace, the current marketplace this id set is generated for.
+    When
+        - checking if it should be part of the mpV2 id set.
+    Then
+        - return True since this item should be skipped.
+        - don't update the excluded item dict since we only want to add content items and not packs
+
+    """
+    import demisto_sdk.commands.common.tools as tools
+    excluded_items_dict = {}
+    mocker.patch.object(tools, 'get_mp_types_from_metadata_by_item', return_value=['xsoar'])
+    pack_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackXsoarMPOnly', 'pack_metadata.json')
+    res1 = should_skip_item_by_mp(pack_path, 'mpv2', excluded_items_dict)
+    assert res1
+    assert not excluded_items_dict
+
+
+def test_should_skip_item_by_mp_update_excluded_dict(mocker):
+    """
+    Given
+        - path of a script which related only to xsoar, the current marketplace this id set is generated for.
+    When
+        - checking if it should be part of the mpV2 id set.
+    Then
+        - return True since this item should be skipped.
+        - update the excluded item dict since we want to add content items to the dict
+
+    """
+    import demisto_sdk.commands.common.tools as tools
+    excluded_items_dict = {}
+    mocker.patch.object(tools, 'get_mp_types_from_metadata_by_item', return_value=['xsoar'])
+    script_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackScriptIsXsoarOnly', 'Scripts', 'DummyScript.yml')
+    res2 = should_skip_item_by_mp(script_path, 'mpv2', excluded_items_dict)
+    assert excluded_items_dict
+    assert res2
+
+
+def test_add_item_to_exclusion_dict():
+    """
+    Given
+        - path of content item, the current marketplace this id set is generated for.
+    When
+        - when creating the id set, checking a content item that is 'xsoar only' if it should be part of the mpV2 id set.
+    Then
+        - return True since this item should be skipped.
+    """
+
+    expected_result = {'CortexXDR': {('integration', 'Cortex XDR')}}
+    excluded_items_from_id_set = {}
+    file_path = os.path.join(TESTS_DIR, 'test_files', 'Packs', 'CortexXDR', 'Integrations', 'PaloAltoNetworks_XDR',
+                             'PaloAltoNetworks_XDR.yml')
+    add_item_to_exclusion_dict(excluded_items_from_id_set, file_path, "Cortex XDR")
+
+    assert IsEqualFunctions.is_dicts_equal(expected_result, excluded_items_from_id_set)